Updated openstack/openstack Project: openstack/tripleo-heat-templates  cec6d0d3dd547c40d66bbc594d7cd7ae037c6214
Change the directory for httpd certs/keys to be service-specific

This moves the directories containing the certs/keys for httpd one step
further inside the hierarchy. This way we will be able to bind-mount
this certificate into the container without bind-mounting any other
certs/keys from other services.

bp tls-via-certmonger-containers

Change-Id: Ibe6e66ae4589b9eab7db330dd8b178e0f8775639
Depends-On: I0b71902358b754fa8bd7fdbb213479503c87aa46