Updated openstack/openstack Project: openstack-infra/zuul  0a38b6c5680f5c7cb23fab91dbfc84f88024829d
Handle dict extras in Voluptuous Schema

Not sure if schema requirements changed but it is no longer valid to
use:

  Schema({}, extra=True)

The docs say that should be avoided and instead you should use:

  Schema(dict)

To indicate a dict that can have any key: value pairs. Additionally
extra=True appears to be replaced with extra=ALLOW_EXTRA so update that
in cases where we have some subset of desired keys.

Change-Id: Ia9f4d535b9b633e703465f155c4722d9a58c45a7