Updated openstack/openstack Project: openstack/glance  9a9faf0f29e3f623d907c24593b0d403be5c2483
Document the duties of the Release CPL

I've written some documentation around the role and duties of the
release CPL due to the fact that I'm leaving this role, Glance, and the
larger OpenStack community. This should provide context for both new and
old members of the Glance community as to their responsibilities and
what is expected of them. It also passes on some experience-driven
advice as to what to do to ensure a better Glance release for all.

Co-Authored-By: Hemanth Makkapati <hemanth.makkapati@rackspace.com>
Change-Id: I8d580d05668d1560b027736fbc1fb1dd36edbabe