Updated openstack/openstack Project: openstack/horizon  359467b4013bb4f89a6a1309e6eda89459288986
Retrieve quota and usage only for resources really required

tenant_quota_usage() is used to retrieve quota and usage
to determine if a resource can be created.
However, tenant_quota_usage retrieves quota and usage for
all resources and it can be a performance problem.

This commit allows to load quota and usage only for resources
which are actually required.

Closes-Bug: #1675504
Change-Id: Iab7322a337a451a1a040cc2f4b55cc319b1ffc4c