Updated openstack/openstack Project: openstack-infra/system-config  536636402efebe97493b4d2ad7520da06e8fe6fd
Remove puppet pinning logic

During the transition from puppet 2.7 to puppet 3.x we needed to
add logic to carefully select what versions of puppet and related
packages would be installed from the puppetlabs apt and yum
repositories. Before puppet 4 was announced, we rightfully feared
accidentally upgrading to backwards-incompatible versions of these
packages and breaking the world.

Now we are completely off of puppet 2.7, so we don't need to keep any
remnants of that around. Moreover, the new packaging system for puppet 4
will make it impossible to accidentally upgrade[1]. The package name has
changed from "puppet" to "puppet-agent", which has facter and hiera etc.
bundled into the same package, and in order to get the new package we
must add a new "puppet collections" repository. So, not only is this
pinning logic not needed to keep us safe, it is also not going to be
useful when we upgrade to puppet 4.

Looking at the puppetlabs repositories[2][3] we are already using the
latest version of the packages we are pinning and Puppet is definitely
not going to add new versions to these repositories.

The $puppet_version variable was leftover from when puppet.conf was
managed by puppet and not ansible so this patch cleans that up as well.

[1] https://docs.puppet.com/puppet/4.9/about_agent.html
[2] http://apt.puppetlabs.com/dists/trusty/main
[3] http://yum.puppetlabs.com/el/7/products/x86_64/

Change-Id: I06b5cd87ee7816b0f929d5e64dc66a5cceca222a