Updated openstack/openstack Project: openstack/nova-specs  01d9b9dacc37ac6320312c2cf8b70cbf3ebb4981
Deprecate the os-hosts API (spec)

Much of the os-hosts API is duplicated with the os-services and
os-hypervisors APIs. Other parts are not implemented for all backends, or
even a good idea for Nova to have in the compute API. Other parts are not even
implemented for any backend anywhere. This spec proposes to deprecate the API
altogether.

APIImpact

Spec for blueprint deprecate-os-hosts

Change-Id: I59ec2bf93b92c84d16bfd7c1ff3e172343c22310