Updated openstack/openstack Project: openstack/neutron  598428fb2e1b6032475db706f1e4325348a97ce2
execute: don't call greenthread.sleep directly

The module may be used in scope that is not using eventlet (f.e. from
inside neutron-keepalived-state-change), in which case a call to
greenthread.sleep from the main thread will result in AssertionError.

Instead of that, call time.sleep that will redirect to greenthread.sleep
in eventlet environment.

This reflects what we do in oslo.concurrency execute implementation.

Change-Id: Id73c0e5015feca2c559da4a68a70aeaaf5cec2a7
Closes-Bug: #1680183