Updated openstack/openstack Project: openstack/heat  98e6c7f04d04243943692538503d4ff070c0371e
Deprecate 'attachments' in favor of 'attachments_list'

The 'attachments' attribute of OS::Cinder::Volume is being processed by
Heat as a string rather than the native list of dicts that come from
Cinder. This makes it hard to consume the data downstream.

This change marks the 'attachments' attribute deprecated in the pike
release in favor of 'attachments_list', which has the correct type of LIST.

An example of where this change is useful is in python-openstackclient;
using the new attribute changes the output of
'openstack stack output show mystack' from:

  ...
  output_value:
    - u"[{u'server_id': u'0f5731c1-da17-4209-a2ef-270c7056f9a3', ... }]"
  ...

to:

  ...
  output_value:
    - attached_at: '2017-03-31T14:05:28.000000'
    - attachment_id: 19436dc5-233d-49cc-a719-f6a92bff466c
    ...
  ...

Change-Id: I52746f87a3872b18e7ae9a7296d6abd2c12c8b9a
Closes-Bug: #1679087