Updated openstack/openstack Project: openstack/nova-specs  7add23514cffb4698d3e381d9c1cd086e40e6dff
Mitigate 'no "man_pages" config value' error

For some reason, the gate was failing with the following message:

    "WARNING: no "man_pages" config value found; no manual pages will be
    written"

I don't know why this has suddenly started happening, but it was
mitigated by 'I97d4ade4'. However, this change didn't fix the root cause
of the issue, which is that 'pbr' defaults to building both html and man
page output and we don't need/want the latter.

Disable the man page builder entirely. This allows us to remove all man
page-related configuration in the process, seeing as it'll never be
used.

Change-Id: I83bb54f413cbad29afb2233043150fb037bca3bd