Updated openstack/openstack Project: openstack/neutron  84901afefc7d316608a59b990d61e138070a437f
DVR: Don't clean snat-ns of DVR HA router when fullsync

Since [1], when the l3 agent does fullsync, for every router, it calls
ensure_snat_cleanup depending on whether the agent is dvr_snat or not.
However, DVR+HA routers always have snat namespaces on dvr_snat agents
holding themselves for keepalived. Therefore, the cleanup call is
unexpected and will cause a series of issues.

This patch ensures that snat namespaces of DVR+HA routers will not be
cleaned when the agent do fullsync.

[1] https://review.openstack.org/#/c/326729/

Change-Id: I5df0a1404f1a80ab0b226d7a60c2885e24247e02
Closes-Bug: #1632540