Updated openstack/openstack Project: openstack/nova  9e071f466b2e3599d78ce0a8895a03c68897ef6f
Add uuid to Service model

This change adds a uuid column to the Service model and versioned
object.  This is a prerequisite to replacing id with uuid in the
services API.

The uuid is generated automatically for new services.  For existing
records, it is generated on read from the database.

Change-Id: I85c39ab422b9687bf032cbe12d17a2babbda1c07
Partially-Implements: blueprint service-hyper-uuid-in-api