Updated openstack/openstack Project: openstack/neutron-specs  22b8477ac8dfd9c1282405fa3afa03c4be35b757
Fix doc build

pbr 2.1.0 has a regression that it always tries to build 'man'
and this causes 'WARNING: "table cell spanning" not supported'.
We specify warning-is-error = 1 and as a result all documentation build
now fails.

This can be fixed by excluding pbr 2.1.0 in requirements.txt.
Note that pbr 2.1.0 is already marked as blacklist in global-requirements.txt.

'man_pages' config option is also dropped from doc/source/conf.py
as we are not interested in manpages as output at all.

Related-Bug: #1681983

Change-Id: Id9aa7b128ad8d4af5db6b96a758f1134a774db1b