Updated openstack/openstack Project: openstack/neutron  1cf557760366d3890a4aa7b3f1779813896b45b6
DocFix: sriov_nic_agent supports qos

The sriov_nic_agent added qos support with the following
patches[1][2][3], and the document was not updated to reflect that.

The bandwidth limit rule was added by [1]
The minimum bandwidth rule was added by [2] and [3]

This commit removes QoS from the "to be supported in the future"
part of the body and adds a partially supported part to the body.

[1] I364fc8158e502d4dcc3510d6157f12969961a11d
[2] Iefe5e698e99d186202d6ef170f84e93bfbba46dd
[3] I6b619a96a2bfde164646c71409b671352bc6ce7d

Change-Id: Ia5e90307795683679a92214786e8d8021eb9ff57