Updated openstack/openstack Project: openstack/neutron  19eb982d58454501a6fa1d553606b53b7a324efd
Floating IP association without subnet gateway IP

Neutron throws an exception when we try to attach the floating ip
to the instances that are launched in a network whose subnet is
connected to the external router by an Ip address otherthan its
gateway-ip. Now the neutron will execute the request with no
exception/warning.

Closes-Bug: #1515990

Change-Id: If212c36d918ed57400a53f4b5fa1925b3d1fa6fd