Updated openstack/openstack Project: openstack-infra/shade  f2f036f0de490fde56d47ccca73017009d391183
Stop defaulting container_format to ovf for vhd

It turns out a basic vhd file is in bare format, not ovf - and bare is
a perfectly sensible thing for vhd files to be.

Actually, even past that, it turns out that container_format is
essentially ignored by glance, not being very useful. disk_format is the
important bit for external consumers.

Also, we grab disk_format twice.

Change-Id: Idc90a8542e1351da641e42aa67d8fa1c37b544e1