Updated openstack/openstack Project: openstack-infra/devstack-gate  ce56c6704ed9fdd733ceef185557c70a0b2caaaf
Simplify network sanity checks

Simplify the network sanity check using the new /etc/ci/mirror_info.sh
script. If thats present we use it to check pings to the mirror host and
HTTP against the pypi mirror. If it isn't present we don't bother. Note
this makes the sanity check fail the job again if it doesn't work which
should be fine at this point because /etc/ci/mirror_info.sh should only
exist if you have a mirror to provide information about.

Change-Id: Ifddb7c5b7ae3e446005338c9679dd5441bc57062