Updated openstack/openstack Project: openstack/tempest  0e50df450b83db91b8b46072b590bf4b260ed2e7
TestSecurityGroupsBasicOps: log console from access point instance

If a test case fails, we log console output for all servers registered
in tenant.servers. So far we have not included the access point servers
for alt_tenant in the log, so in case of failure we miss a crucial data
point: what happened inside the "access point" instance itself.

Change-Id: I52594fb6d2a86abeb2e00c2ff24b3f7d6042222d