Updated openstack/openstack Project: openstack-infra/system-config  a6b22d066d87828d4421f6a78aa318e8835dd7ed
Move pip management to openstack_project::server

First we remove the `include ::pip` from single_use_slave, since it was
already being declared in openstack_project::template and it is a fluke
of puppet that it wasn't causing a duplicate resource error.

Then we move the pip puppet class and the virtualenv package resource
out of openstack_project::template to openstack_project::server. The
virtualenv package should already installed on nodepool workers by DIB.
The pip puppet class does three things:

  1. Installs the python devel package, which we are already getting
     from DIB
  2. Installs pip, which we are already getting from DIB
  3. Manages pip.conf. Here in the puppet manifests we're setting the
     index URL to the upstream pypi repository, not our mirrors. When
     the node is launched in nodepool the configure_mirror.sh ready
     script will undo this and configure pip to use our mirrors.

So there should be no need to use the pip class on the nodepool workers.

Change-Id: Icc720e61cd12f31113a6e12482d4903a6772ae69