Updated openstack/openstack Project: openstack-infra/project-config  a692b93caae4f8945c8d1d56ddfbbe816b6e1edc
Introduce read-only ACL and start using it

instead of creating a separate ACL file for each repo that is read-only,
create a new read-only config file and use that. This avoids
duplications and makes it easier to mark repos read-only.

This change covers only the first few repos, further repos will be done
in a followup.

Change-Id: I7264070a085eb999167e050016fee37eab1770b7