Updated openstack/openstack Project: openstack-infra/shade  469eef95e72acd53dd05f5c4cacd5b1b2e5fbb33
Use keystone_session in _get_raw_client

We're bypassing our local session object when we construct the adapter
and instead getting a raw one from occ. This becomes aparent when we try
to do things to our local one and they don't show up in the one used by
the adapter.

Change-Id: I91673527ebe1e1446dfc6efb8c495694a4a65c6b