Updated openstack/openstack Project: openstack/neutron  81de3606bbaa2827f39b425d9ea46d4e071a78e7
Refactoring _create_test methods for OVO UTs

Most of the _create_test methods create a property and only uses
its id during their execution.  This change pretends to simplify
the returned data and reduce the number of lines used for the
creation of test data used during the execution of Unit Tests for
Oslo-Versioned Objects.

Change-Id: Icfc251a7d460246e91e260a4ac503d10c4031170
Partially-Implements: blueprint adopt-oslo-versioned-objects-for-db