Updated openstack/openstack Project: openstack/neutron  b8ab4c9ccecb68050a0844442fe8a1dade52e497
Add IPAllocations to the port fixed_ips

Since the IPAM is currently inserting the IPAllocation objects directly
into the DB, the make_port_dict method triggers queries to the DB when
the fixed_ips relationship is accessed, which negatively affects the
performance of Neutron.

This patchset changes this behavior, by appending IPAllocations to the
port's fixed_ips relationship. This change will facilitate the
implementation of enginefacade

Change-Id: Ib32509d974c8654131112234bcf19d6eae8f7cca
Closes-Bug: #1676235