Updated openstack/openstack Project: openstack/neutron  e413e81508cc56862c6eb15b98a9e1500c00b4c8
Move notify_security_groups_member_updated to callback

This moves the notify_security_groups_member_updated calls
outside of ML2 into a callback in the security groups RPC
base.

This will allow us to ultimately stop overriding the
create/delete/update port inheritance in ML2.

This results in a small behavior change internally for bulk
port creation operations where the notifier will be called
once for each port rather than once for all ports. We can
revisit this in the future if it turns out to be problematic
but right now the bulk port creation API is rarely anyway
since it's mostly a hidden feature (lack of client support).

Partial-Bug: #1666424
Change-Id: I344a82d5a1a8bf1e2b3f6a3ce7725682bb03b007