Updated openstack/openstack Project: openstack/neutron  4bffb7defb32ca52125bebe20c8f0ac848f54d13
Don't load rels on detached objects

OVO detaches DB objects from the session and then holds onto
them for the life time of the OVO object. Simultaneously we attach
a weakref to the object to the session to load relationships for it
on commit.

If GC hasn't eliminated this object by the time we try to load
relationships, we end up trying to load rels for a detached object,
which implicitly tries to add it back into the session. This fails
if there is another copy in the session.

This patch just skips loading detached objects because we don't need
relationships loaded for them anyway.

Closes-Bug: #1683469
Change-Id: Ida3272562edb6608c6bfd7e11267159d6565a0c7