Updated openstack/openstack Project: openstack-infra/project-config  9c6aeb7528828e6357a5a046315637072f41dafd
Add os_karbor to OpenStack-Ansible

This patch adds a new empty git repository to OpenStack-Ansible for
deploying Karbor.

This project will have its own core team with the general
openstack-ansible core team included, with the appropriate rights to
ensure that they can act as a back-stop in case of any issues.

Change-Id: Ia73f41427ab9360db057e1ce710f7b92375e0d3d
Needed-By: I465da438762f3b675febd10009270ba6add8f27f