Updated openstack/openstack Project: openstack/nova  a2d8680cc63abdd4e9e3a960fc7019f9f032a4f0
Remove aggregate uuid generation on load from DB

As the comments in the code suggest, we can remove this
code now. There was a blocking database migration which
enforced that the online data migrations to generate
uuids for aggregates has been completed:

4d0915568a1011ddee7317ddfb237be0803e7790

So this code is safe to remove now. As a result, we
also remove the online data migration routine which
relies on the object code.

Change-Id: I2050b5bdf906d2d2f8a87c79ca0804e0fc955755