Updated openstack/openstack Project: openstack/nova  a41136dfee426a3e99b30c5ed0de246c67e4bada
Deprecate CONF.api.allow_instance_snapshots

The introduction of this option predates Gerrit [1]. It is
not used consistently, since it's not used for createBackup
or shelve APIs which also create snapshots. And it's not
discoverable.

We need to deprecate this option and anyone that cares to
disable snapshots needs to do it via policy.

[1] https://github.com/openstack/nova/commit/9633e9877c7836c18c30b51c8494abfb025e64ca

Change-Id: I321c04376d86ddb1bb3c8f1fa938a7b424e658d0