Updated openstack/openstack Project: openstack/ironic  b0c63477d1c0fe6a49fa8a906bc224678789b476
Make config generator aware of 'default_log_levels' override

The way it is done in this patch is recommended by the oslo.config
documentation in [0], so that the updated default_log_levels value
is correctly reflected in the generated ironic.conf.sample file.
An oslo.config.opts.defaults entry point is added.

[0] https://docs.openstack.org/developer/oslo.config/generator.html

Change-Id: Ic200e8d0d721a6c94bd17068a6e5000db930eefd