Updated openstack/openstack Project: openstack/python-novaclient  90a14a680543cbd8f33422584145e7eae1ce7c4b
Explicitly specify arguments of server_groups creation

Create API call expectes two parameters - name and policies[*]. It is not a
complex structure, so there is no reason to support kwargs. It is better to
specify explicitly all expected parameters.

PS: removing support of kwargs is "backward compatible", since we have
validation at API side and usage of additional parameters was an error
previously.

[*] - https://developer.openstack.org/api-ref/compute/?expanded=create-server-group-detail

Change-Id: I8b40b0db450287bbc5ee8d69834aa764353dbd98