Updated openstack/openstack Project: openstack/cinder  19d07a1f394c905c23f109c1888c019da830b49e
Check "kernel_id", "ramdisk_id" for Glance v2 only

kernel_id and ramdisk_id are stored as the image's standalone fields
with Glance V2. But they are contained in the images'properties already
with Glance V1.

After the change[1], the "_extract_attributes" is used for Glance V1 only.
So if use Glance V1, Cinder don't need to get them again.

The test cases is covered by the change[1] already.

[1]: Ib54bb6759b27334294fb2c6d2c0bfe4eae3d0920

Closes-bug: #1669662
Change-Id: Ifa32fefa29c6629cd7f7aa40eaf8b0df4828c413