Updated openstack/openstack Project: openstack/neutron  1ae2c9a2f60b2ae4e22b585bd3a289f20c68aee1
Update auto-addresses on MAC change

On MAC update add an entry for auto-addressed IPs so
that they get removed and re-allocated.

Also Fix the unit test that was testing for this, so that
it checks the IP for the new updated port (it was checking
the port before the update).

Change-Id: I0a83efa7b8dfa1f6979566846195fb349547018a
Closes-Bug: #1671548