Updated openstack/openstack Project: openstack/neutron  f4da36fa2c390fb401a62f135c383dfbf2f905b9
Switch to neutron-lib hacking factory

This allows us to kill a bunch of in-tree checks. There are still some
checks that are either not yet in the library, or don't belong there in
the first place, so we still stick to our own factory, just reusing
whatever is there in neutron-lib.

This change skips some new checks. We gotta figure out what to do with
those, that will belong in a separate follow-up.

Change-Id: Ifeb40ec0e0c4ca623b33a6b9f500dec15cec4de0