Updated openstack/openstack Project: openstack/neutron  73f78a02032c2d8e5143fe3abb3433efc8b9ebe8
Stop loading OVOServerRpcInterface in ML2 testcase

This RPC notifier has an asynchronous handler that does DB lookups
to send notifications after an object has been created. Not only
is this a waste of resources in unit tests, it interferes with tests
that are inspecting the DB queries that occur between calls.

This just blocks the OVOServerRpcInterface in the ML2 test case and
then enables it for the set of unit tests that needs it.

Change-Id: I807e2445da9be241c24d8b6e5bda8b12927aa7fc
Closes-Bug: #1679775