Updated openstack/openstack Project: openstack/governance  99e92086113f683a177c46cfac8ee5f923ce2abe
Propose the addition of a status:maintenance-mode

The change proposes the addition of a status that can be applied to a
project to indicate that the project has entered a transient period of
low activity and contribution.

The change follows the tag-template and provides a rationale, specific
conditions that must be met by the project, rules for adding and
removing the tag, and deprecation rules.

Change-Id: I471504c231a50a9bcd0ddb49b3d7534e5af6cc27
Depends-On: Ib032885003b17e70bf3fc9768ffc4383b834d2ff