Updated openstack/openstack Project: openstack/governance  45d78042dd595781da9a7467804dacaf791f6ea7
Remove App Catalog from official projects

With our project structure it is easy for new promising projects
to be recognized as an official OpenStack project (produced by
the OpenStack community). The other side of that coin is the need
for us to recognize when a project fails to attract its desired
target, or when the landscape around us changed to the point where
a given project may no longer make sense.

The App Catalog was created early 2015 as a marketplace of
pre-packaged applications that you can deploy using Murano.
Initially a demo by Mirantis, it was very successfully converted
into an open upstream project team, and deployed as a "beta" as
apps.openstack.org.

Since then it grew additional categories (Glance images, Heat &
Tosca templates), but otherwise did not pick up a lot of steam.
The website (still labeled "beta") features 45 glance images,
6 Tosca templates, 13 heat templates and 94 murano packages
(~30% of which are just thin wrappers around Docker containers).
Traffic stats show around 100 visits per week, 75% of which only
read the index page.

In parallel, containers emerged as an application packaging format,
and Docker developed a pretty successful containerized applications
marketplace (the Docker Hub), with hundreds of thousands of
regularly-updated apps. Helm is showing great promises of taking
those base containers to the next level and allow to conveniently
deploy more complex applications.

An application marketplace value is directly linked to how many
applications you can find in it, and how regularly-updated they
are. In the age of containers, keeping the App Catalog around
make us look like we are unsuccessfully trying to compete with
that ecosystem, while OpenStack is in fact completely complementary.
Our focus should be on co-opting that ecosystem, by making it dead
simple to start any container or helm chart on resources provided
by OpenStack infrastructure (including work on Zun and Magnum).

I would therefore move to make the Community App Catalog an
unofficial project, and no longer expose it on apps.openstack.org.
It is not an easy decision, especially as the community team there
did great technical work, is composed of personal friends and is
still active. But it is a decision which I think we need to take,
as at this point keeping it around is hurting our achievement of
the OpenStack mission more than it helps it.

Change-Id: Ie3977a685bc28275d2c109d47b8189c52ed11eda