Updated openstack/openstack Project: openstack/training-guides  50551f534653369f2f56200378401959f2f8bcc1
Add Ref slide for howitsmade-official-projects

There are many links of reference sites/topic used in slide
as note (which are only viewable in presenter view).

Those links are helpful for audience to get further details
and good to mention in reference slide.

This commit add new reference slide with related links.

Change-Id: I2958d524e71b8c86362ec2a8ee3fb9c99f094b1e